Remove logic and files to simplify creation
- Remove docker file creation - irrelevant for initial setup
- Remove code that is turned into 'backend' files
- Add '.env.local' to 'setup_assets' root
- Sanitise 'filepaths.py' constants
- Refactor 'StaticAssetsController' to single task to simplify
