+++ create_api_app/setup_assets/.env.local
@@ -1,0 +1,32 @@
+# An optional API URL only used when populating your database with data from an existing API. Refer to 'backend/db_insert.py'
+API_URL=
+
+# MongoDB: The connection url to your database
+# https://www.mongodb.com/
+DB_URL=
+
+# Uploadthing: storing files and handling file uploading
+# https://uploadthing.com/
+UPLOADTHING_SECRET=
+NEXT_PUBLIC_UPLOADTHING_APP_ID=
+
+# Clerk: User Authentication
+# https://clerk.com/docs/quickstarts/nextjs
+NEXT_PUBLIC_CLERK_PUBLISHABLE_KEY=
+CLERK_SECRET_KEY=
+
+NEXT_PUBLIC_CLERK_SIGN_IN_URL=/auth/sign-in
+NEXT_PUBLIC_CLERK_SIGN_UP_URL=/auth/sign-up
+NEXT_PUBLIC_CLERK_AFTER_SIGN_IN_URL=/
+NEXT_PUBLIC_CLERK_AFTER_SIGN_UP_URL=/
+
+# Stripe: user payments
+# https://stripe.com/docs
+NEXT_PUBLIC_STRIPE_PUBLISHABLE_KEY=
+STRIPE_SECRET_KEY=
+STRIPE_WEBHOOK_SECRET=
+NEXT_PUBLIC_STRIPE_CLIENT_ID=
+NEXT_PUBLIC_PLATFORM_SUBSCRIPTION_PERCENT=1
+NEXT_PUBLIC_PLATFORM_ONETIME_FEE=2
+NEXT_PUBLIC_PLATFORM_PERCENT=1
+NEXT_PRODUCT_ID=
